% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontcm.r
\docType{package}
\name{fontcm-package}
\alias{fontcm}
\alias{fontcm-package}
\title{fontcm: Computer Modern Font for Use with Extrafont Package}
\description{
Computer Modern font with Paul Murrell's symbol extensions. Is is to be used with the **extrafont** package. When this font package is installed, the CM fonts will be available for PDF or Postscript output files; however, this will (probably) not make the font available for screen or bitmap output files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fbertran/fontcm}
  \item Report bugs at \url{https://github.com/fbertran/fontcm/issues}
}

}
\author{
\strong{Maintainer}: Frederic Bertrand \email{frederic.bertrand@lecnam.net} (\href{https://orcid.org/0000-0002-0837-8281}{ORCID})

Authors:
\itemize{
  \item Winston Chang
  \item Alexej Kryukov
  \item Paul Murrell
}

}
\keyword{internal}
