% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{fontcm}
\alias{fontcm}
\alias{fontcm-package}
\alias{package-fontcm}
\title{Computer Modern font package}
\description{
This package contains the Computer Modern font with Paul Murrell's
symbol extensions. It is meant to be used with the fonts package.
}

